+++ R/get_comparisons.R
@@ -17,8 +17,6 @@ get_comparisons <- function(
     # get_se_delta() needs perturbed coefficients model
     model <- if (is.null(model_perturbed)) mfx@model else model_perturbed
 
-    settings_init()
-
     predictions <- predictions_hi_lo(model, lo, hi, type, ...)
     list2env(predictions, environment())
 
@@ -130,6 +128,7 @@ get_comparisons <- function(
 
     out[, predicted_lo := pred_lo[["estimate"]]]
     out[, predicted_hi := pred_hi[["estimate"]]]
+
 
     if (!is.null(pred_or)) {
         out[, predicted := pred_or[["estimate"]]]