+++ R/methods_brms.R
@@ -94,7 +94,13 @@ get_predict.brmsfit <- function(
         out <- apply(draws, c(2, 3), stats::median)
         levnames <- dimnames(draws)[[3]]
         if (is.null(levnames)) {
-            colnames(out) <- seq_len(ncol(out))
+            cols <- tryCatch(levels(insight::get_response(m)),
+                error = function(e) NULL)
+            if (is.null(cols)) {
+                colnames(out) <- seq_len(ncol(out))
+            } else {
+                colnames(out) <- cols
+            }
         } else {
             colnames(out) <- levnames
         }