+++ R/package.R
@@ -1,8 +1,6 @@
 # https://stackoverflow.com/questions/9439256/how-can-i-handle-r-cmd-check-no-visible-binding-for-global-variable-notes-when#comment20826625_12429344
 # 2012 hadley says "globalVariables is a hideous hack and I will never use it"
 # 2014 hadley updates his own answer with globalVariables as one of "two solutions"
-#' @importFrom Rcpp evalCpp
-#' @useDynLib marginaleffects
 #' @import data.table
 utils::globalVariables(c(
     ".",
@@ -11,7 +9,6 @@ utils::globalVariables(c(
     "..cl",
     "..col_newdata",
     "..col_x",
-    "..keepers",
     "..cols",
     "..hypothesis_by",
     "..idx",
@@ -21,6 +18,7 @@ utils::globalVariables(c(
     "..idx_hi",
     "..idx_lo",
     "..idx_or",
+    "..keepers",
     "..lab",
     "..lab_cols",
     "..label_columns",
@@ -55,6 +53,7 @@ utils::globalVariables(c(
     "lo",
     "logLik",
     "m",
+    "mAD",
     "marginaleffects_contrast_hi",
     "marginaleffects_contrast_lo",
     "marginaleffects_function",