+++ NEWS.md
@@ -1,5 +1,20 @@
 # News {.unnumbered}
 
+## 0.30.0
+
+New:
+
+* Automatic differentiation provides much faster and more accurate standard errors for several model types. Thanks to @t-kalinowski for improvements to `reticulate` handling and to @arcruz0 for `JAX` prediction functions.
+* `sampleSelection` models with `type="conditional"` are now supported.
+* Support fitting functions from the `partykit` package: `lmtree` and `glmtree`.
+
+Bug:
+
+* `avg_predictions()` did not respect the `wts` argument when using `variables` to create counterfactual predictions. Thanks to @strengejacke for report #1596.
+* Binary variables treated as integers by `datagrid()` with `newdata=` instead of `model=`. Issue #1584.
+* Better labelling in some `brms` models with categorical outcomes. Thanks to @mattansb for a proposed solution and to @strengejacke for report #1392.
+* Improve printing of grouped hypotheses. Issue #1579.
+
 ## 0.29.0
 
 New functions: